ble_sensor_mqtt_gw/requirements.txt
jsonschema==2.6.0
paho-mqtt==1.5.0
